requirements.txt
ruff==0.12.9
pre-commit==4.3.0
django-browser-reload==1.19.0
shot-scraper==1.8
Pillow==12.0.0
